pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "remote-mcp-server"
version = "1.0.0"
description = "A production-ready MCP server for PostgreSQL database access"
authors = [{name = "UnlockMCP Team", email = "team@unlockmcp.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
keywords = ["mcp", "model-context-protocol", "postgresql", "ai", "database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "psycopg2-binary>=2.9.7",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Unlock-MCP/remote-mcp-server"
Documentation = "https://unlockmcp.com/guides/how-to-build-a-remote-mcp-server"
Repository = "https://github.com/Unlock-MCP/remote-mcp-server"
Issues = "https://github.com/Unlock-MCP/remote-mcp-server/issues"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]